ia-analyzer/requirements.txt
opencv-python-headless==4.8.1.78
numpy==1.26.4
requests==2.31.0
schedule==1.2.1
ultralytics==8.1.34
sympy==1.12
python-dotenv==1.0.1
